chore: Rename BuyButton_temp to BuyButtonShop
Renamed the BuyButton_temp component to BuyButtonShop and updated its import in the product page for clarity and consistency.
