+++ app/products/[productId]/page.tsx
@@ -1,5 +1,5 @@
 import BackButton from "@/components/BackButton";
-import BuyButton from "@/components/BuyButton_temp";
+import BuyButton from "@/components/BuyButtonShop";
 
 import Image from "next/image";
 import { notFound } from "next/navigation";
